SingleSessionView with polymer template. Initial state.
- Running successful
- Tests passed
- Build successful
